+++ CHANGELOG.md
@@ -12,6 +12,8 @@ We use *breaking :warning:* to mark changes that are not backward compatible (re
 
 It is recommend to upgrade the storage components first (Receive, Store, etc.) and then Queriers. This will enable batching support. Otherwise, you risk high memory usage in the Querier component if gRPC compression is enabled.
 
+- [#8594](https://github.com/thanos-io/thanos/pull/8594): Query: Support per endpoint TLS configuration
+
 ### Fixed
 
 ### Added